+++ src/clj/instaparse/viz.clj
@@ -25,8 +25,8 @@
     (r/tree->image sequential? rest mytree 
                  :node->descriptor (fn [n] {:label (if (sequential? n) 
                                                      (apply str (first n)
-                                                            (when (span n) 
-                                                              ["\n" (span n)]))
+                                                            (when (span n)
+                                                              ["\\n" (span n)]))
                                                      (with-out-str (pr n)))})
                  :options options))
       
@@ -37,8 +37,8 @@
              :node->descriptor (fn [n] 
                                  {:label (if (and (map? n) (:tag n))
                                            (apply str (:tag n)
-                                                  (when (span n) 
-                                                    ["\n" (span n)]))
+                                                  (when (span n)
+                                                    ["\\n" (span n)]))
                                            (with-out-str (pr n)))})
              :options options))
 
@@ -73,16 +73,18 @@ to your dependencies, and installed graphviz on your system.
 See https://github.com/ztellman/rhizome for more information."
   [tree & {output-file :output-file options :options}]
   {:pre [(not= (tree-type tree) :invalid)]}
-  (let [image
-        (try
-          (case (tree-type tree)
-            :enlive (enlive-tree-viz tree options)
-            (:hiccup :nil) (hiccup-tree-viz tree options)
-            :rootless (tree-viz (fake-root tree) options))
-          (catch IOException e
-            (throw (UnsupportedOperationException. 
-                     "\n\nYou appear to have rhizome in your dependencies, but have not installed GraphViz on your system.
+  (let [ttype (tree-type tree)]
+    (if (= ttype :rootless)
+      (tree-viz (fake-root tree) :options options)
+      (let [image
+            (try
+              (case (tree-type tree)
+                :enlive (enlive-tree-viz tree options)
+                (:hiccup :nil) (hiccup-tree-viz tree options))
+              (catch IOException e
+                (throw (UnsupportedOperationException. 
+                         "\n\nYou appear to have rhizome in your dependencies, but have not installed GraphViz on your system.
 \nSee https://github.com/ztellman/rhizome for more information.\n"))))]
-    (if output-file 
-      (r/save-image image output-file)
-      (r/view-image image))))
+        (if output-file 
+          (r/save-image image output-file)
+          (r/view-image image))))))