+++ src/cljs/instaparse/core.cljs
@@ -24,7 +24,7 @@
   {:pre [(#{:abnf :ebnf} type)]}
   (set! *default-input-format* type))
 
-(declare failure?)
+(declare failure? standard-whitespace-parsers)
 
 (defn- unhide-parser [parser unhide]
   (case unhide
@@ -110,58 +110,90 @@
       
       :else
       (gll/parses (:grammar parser) start-production text partial?))))
-  
+ 
 (defrecord Parser [grammar start-production output-format]
   IFn
-  (-invoke [parser text] (parse parser text))
-  (-invoke [parser text key1 val1] (parse parser text key1 val1))
-  (-invoke [parser text key1 val1 key2 val2] (parse parser text key1 val1 key2 val2))
-  (-invoke [parser text key1 val1 key2 val2 key3 val3] (parse parser text key1 val1 key2 val2 key3 val3)))
+  (invoke [parser text] (parse parser text))
+  (invoke [parser text key1 val1] (parse parser text key1 val1))
+  (invoke [parser text key1 val1 key2 val2] (parse parser text key1 val1 key2 val2))
+  (invoke [parser text key1 val1 key2 val2 key3 val3] (parse parser text key1 val1 key2 val2 key3 val3))
+  (applyTo [parser args] (apply parse parser args))) 
 
 (defn parser
   "Takes a string specification of a context-free grammar,
-   or a URI for a text file containing such a specification,
-   or a map of parser combinators and returns a parser for that grammar.
+  or a URI for a text file containing such a specification,
+  or a map of parser combinators and returns a parser for that grammar.
 
-   Optional keyword arguments:
-   :input-format :ebnf
-   or
-   :input-format :abnf
+  Optional keyword arguments:
+  :input-format :ebnf
+  or
+  :input-format :abnf
 
-   :output-format :enlive
-   or
-   :output-format :hiccup
-   
-   :start :keyword (where :keyword is name of starting production rule)"
+  :output-format :enlive
+  or
+  :output-format :hiccup
+
+  :start :keyword (where :keyword is name of starting production rule)
+
+  :string-ci true (treat all string literals as case insensitive)
+
+  :auto-whitespace (:standard or :comma)
+  or
+  :auto-whitespace custom-whitespace-parser"
   [grammar-specification &{:as options}]
   {:pre [(contains? #{:abnf :ebnf nil} (get options :input-format))
-         (contains? #{:enlive :hiccup nil} (get options :output-format))]}
+         (contains? #{:enlive :hiccup nil} (get options :output-format))
+         (let [ws-parser (get options :auto-whitespace)]
+           (or (nil? ws-parser)
+               (contains? standard-whitespace-parsers ws-parser)
+               (and
+                 (map? ws-parser)
+                 (contains? ws-parser :grammar)
+                 (contains? ws-parser :start-production))))]}
   (let [input-format (get options :input-format *default-input-format*)
         build-parser (case input-format 
                        :abnf abnf/build-parser
-                       :ebnf cfg/build-parser)
+                       :ebnf (if (get options :string-ci)
+                               (fn [spec output-format]
+                                 (binding [cfg/*case-insensitive-literals* true]
+                                   (cfg/build-parser spec output-format)))
+                               cfg/build-parser))
         output-format (get options :output-format *default-output-format*)
-        start (get options :start nil)]    
-    (cond
-      (string? grammar-specification)
-      (let [parser (build-parser grammar-specification output-format)]
-        (if start (map->Parser (assoc parser :start-production start))
-          (map->Parser parser)))
-      
-      (map? grammar-specification)
-      (let [parser
-            (cfg/build-parser-from-combinators grammar-specification
-                                               output-format
-                                               start)]
-        (map->Parser parser))
-      
-      (vector? grammar-specification)
-      (let [start (if start start (grammar-specification 0))
-            parser
-            (cfg/build-parser-from-combinators (apply hash-map grammar-specification)
-                                               output-format
-                                               start)]
-        (map->Parser parser)))))
+        start (get options :start nil)
+
+        built-parser
+        (cond
+          (string? grammar-specification)
+          (let [parser (build-parser grammar-specification output-format)]            
+            (if start (map->Parser (assoc parser :start-production start))
+              (map->Parser parser)))
+
+          (map? grammar-specification)
+          (let [parser
+                (cfg/build-parser-from-combinators grammar-specification
+                                                   output-format
+                                                   start)]
+            (map->Parser parser))
+
+          (vector? grammar-specification)
+          (let [start (if start start (grammar-specification 0))
+                parser
+                (cfg/build-parser-from-combinators (apply hash-map grammar-specification)
+                                                   output-format
+                                                   start)]
+            (map->Parser parser)))]
+
+    (let [auto-whitespace (get options :auto-whitespace)
+          ; auto-whitespace is keyword, parser, or nil
+          whitespace-parser (if (keyword? auto-whitespace)
+                              (get standard-whitespace-parsers auto-whitespace)
+                              auto-whitespace)]
+      (if-let [{ws-grammar :grammar ws-start :start-production} whitespace-parser]
+        (assoc built-parser :grammar
+               (c/auto-whitespace (:grammar built-parser) (:start-production built-parser)
+                                  ws-grammar ws-start))
+        built-parser))))
+
         
 (defn failure?
   "Tests whether a parse result is a failure."
@@ -186,3 +218,6 @@
 (def transform t/transform)
 
 (def visualize viz/tree-viz)
+(def ^:private standard-whitespace-parsers
+  {:standard (parser "whitespace = #'\\s+'")
+   :comma (parser "whitespace = #'[,\\s]+'")})